requirements.txt
django
Pillow
psycopg2
djangorestframework
